Organizing, working on logging, updating "controle.avapolos" and adding security to all the services. (#34)
* Working on #15

* Working on #15 part 2.

* Organization.

* Working on #15 part 3.

* Working on #15 part 4.

* Working on #15 part 5.

* Working on #15 part 6.

* Working on #32.

* Working on #32 part 2.

* Reorganizing part 1

* New logger fixing.

* Updating install_networking.sh

* Working on #25

* Fixing errors on data compiling and build.sh

+++ build/avapolos/data/compile.sh
@@ -1,13 +1,9 @@
 #!/usr/bin/env bash
 
-dir=$(pwd)
-cd ../../preinstall
 source header.sh
-cd ../avapolos/scripts/sync/
-source functions.sh
-cd $dir
-unset dir
-source header.sh
+generateSecrets > secrets
+chmod 600 secrets
+source secrets
 
 while true
 do
@@ -25,13 +21,21 @@ do
 			shift
       export KEEPALIVE="true"
 		;;
+		--ignore-base)
+			shift
+      export IGNORE_BASE="true"
+		;;
 		--cleanup)
       for service in $(ls $SERVICES_DIR | grep -v TEMPLATE); do
         cd "$SERVICES_DIR/$service"
         echo "Limpando dados do serviço: $service"
         rm -rf data
-        docker-compose down
+        if [[ -f docker-compose.yml ]]; then
+          docker-compose down
+        fi
       done
+      remove_docker_network avapolos
+      remove_docker_network proxy
       exit 0
 		;;
     -*)
@@ -45,48 +49,45 @@ do
 		esac
 done
 
-echo "Assegurando permissões corretas" | log debug data_compiler
 sudo chown $USER:$USER -R .
 
 mkdir -p $PACK_DIR/data && rm -rf $PACK_DIR/data/*
 
-echo "Criando rede Docker avapolos_build e proxy" | log debug data_compiler
-# add_docker_network avapolos
-# add_docker_network proxy
-docker network create avapolos || true
-docker network create proxy || true
+log debug "Criando rede Docker avapolos e proxy"
+add_docker_network avapolos
+add_docker_network proxy
 
 if [[ -z "$(cat /etc/hosts | grep -o 'AVAPOLOS BUILD START')" ]]; then
-  setHosts | log debug data_compiler
+  log debug "Configurando /etc/hosts"
+  setHosts
 fi
 
-cd $TRAEFIK_DIR
-echo "Compilando traefik" | log info data_compiler
-run compile.sh
-cp -rf $TRAEFIK_DATA_DIR/* "$PACK_DIR/data/"
-
-cd $INICIO_DIR
-echo "Compilando inicio.avapolos" | log info data_compiler
-run compile.sh
-cp -rf $INICIO_DATA_DIR/* "$PACK_DIR/data/"
-
-cd $MANUTENCAO_DIR
-echo "Compilando manutencao" | log info data_compiler
-run compile.sh
-cp -rf $MANUTENCAO_DATA_DIR/* "$PACK_DIR/data/"
+if ! [[ "$IGNORE_BASE" = "true" ]]; then
+  export MANUTENCAO=""
+  BASE_SERVICES="TRAEFIK INICIO MANUTENCAO DHCPD DNSMASQ"
+  log debug "Compilando serviços base: $BASE_SERVICES"
+  for service in $BASE_SERVICES; do
+    dir=$service"_DIR"
+    data_dir=$service"_DATA_DIR"
+    cd "${!dir}"
+    log info "Compilando $service"
+    run compile.sh
+    cp -rf "${!data_dir}"/* "$PACK_DIR/data/"
+  done
+fi
 
 #SERVICES=${SERVICES:-$(ls | grep -v TEMPLATE | grep -v traefik)}
 SERVICES=${SERVICES:-$(for svc in $(cat $ROOT_DIR/../services/enabled_services | grep -v educapes); do echo $(echo $svc | cut -d . -f1); done)}
-echo "Compilando serviços: $SERVICES" | log info data_compiler
+log info "Compilando serviços: $SERVICES"
 for service in $SERVICES; do
-  echo "Compilando serviço $service" | log info data_compiler
+  log info "Compilando serviço $service" 
   cd "$SERVICES_DIR/$service"
   run compile.sh
   cp -rf $SERVICES_DIR/$service/data/* "$PACK_DIR/data/"
 done
 
 if ! [[ "$KEEPALIVE" = "true" ]]; then
-  unsetHosts | log debug data_compiler
+  unsetHosts
   cd $TRAEFIK_DIR
   docker-compose down
   cd $INICIO_DIR
@@ -95,11 +96,15 @@ if ! [[ "$KEEPALIVE" = "true" ]]; then
   docker-compose down
   for service in $SERVICES; do
     cd "$SERVICES_DIR/$service"
-    echo "Parando serviço $service" | log debug data_compiler
+    log debug "Parando serviço $service" 
     docker-compose down
   done
+  remove_docker_network avapolos
+  remove_docker_network proxy
 fi
+
+cp $ROOT_DIR/secrets $PACK_DIR/data
 
 cd $PACK_DIR
 tar --use-compress-program="pigz -9" -cf data.tar.gz data
-cp -rf data.tar.gz $ROOT_DIR/../
+cp -rf data.tar.gz ../
